feat: add chat flow functionality
- Implemented chatFlow in a new file genkit-flows/chatFlow.ts to handle chat interactions.
- Updated index.ts to import and utilize the new chatFlow for the /chat endpoint.
